refactor : Quantity 리팩토링
