+++ src/test/java/kitchenpos/global/vo/QuantityTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.vo;
+package kitchenpos.global.vo;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -8,7 +8,7 @@ import org.junit.jupiter.params.provider.ValueSource;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-@DisplayName("메뉴 CONTEXT 수량")
+@DisplayName("수량")
 class QuantityTest {
 
     @DisplayName("수량 생성")
