+++ src/test/java/kitchenpos/menus/tobe/domain/menu/MenuTest.java
@@ -3,7 +3,7 @@ package kitchenpos.menus.tobe.domain.menu;
 import kitchenpos.menus.tobe.domain.menugroup.MenuGroup;
 import kitchenpos.global.vo.DisplayedName;
 import kitchenpos.global.vo.Price;
-import kitchenpos.menus.tobe.vo.Quantity;
+import kitchenpos.global.vo.Quantity;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
